Slightly changed logic for InputStreams support
Instead of including the InputStream directly in the request, which
causes the faban harness to reply with an error code of 500, we now
read the input stream into an array of bytes, and include it in the
request.

In order to implement this change easily, I added Guava as a dependency
and exploited its `ByteStreams` class.